Refactor Test classes
