+++ tests/acceptance/ProjectType/Incubator/RoboFileCest.php
@@ -3,38 +3,17 @@
 namespace Cheppers\Robo\Drupal\Tests\Acceptance\ProjectType\Incubator;
 
 use Cheppers\Robo\Drupal\Test\AcceptanceTester;
-use Symfony\Component\Filesystem\Filesystem;
+use Cheppers\Robo\Drupal\Tests\Acceptance\Base as BaseCest;
 use Symfony\Component\Finder\Finder;
 use Webmozart\PathUtil\Path;
 
-class RoboFileCest
+class RoboFileCest extends BaseCest
 {
 
     /**
      * @var string
      */
     protected $class = \ProjectTypeIncubatorRoboFile::class;
-
-    /**
-     * @var string[]
-     */
-    protected $tmpDirs = [];
-
-    /**
-     * @var \Symfony\Component\Filesystem\Filesystem
-     */
-    protected $fs;
-
-    public function __construct()
-    {
-        $this->fs = new Filesystem();
-
-        //register_shutdown_function(function () {
-        //    $tmpDirs = array_filter($this->tmpDirs, 'file_exists');
-        //    $this->fs->chmod($tmpDirs, 0700, 0, true);
-        //    $this->fs->remove($tmpDirs);
-        //});
-    }
 
     public function listTest(AcceptanceTester $i): void
     {
@@ -216,34 +195,5 @@ class RoboFileCest
             $i->openFile($filePath);
             $i->canSeeFileContentsEqual($file->getContents());
         }
-    }
-
-    protected function prepareProject(string $projectName): string
-    {
-        $tmpDir = $this->createTmpDir();
-        $templateDir = $this->fixturesRoot($projectName) . '/base';
-        $this->fs->mirror($templateDir, $tmpDir);
-
-        return $tmpDir;
-    }
-
-    protected function fixturesRoot(string $projectName): string
-    {
-        return codecept_data_dir("fixtures/ProjectType/Incubator/$projectName");
-    }
-
-    protected function createTmpDir(): string
-    {
-        $class = explode('\\', __CLASS__);
-        $tmpDir = tempnam(sys_get_temp_dir(), 'robo-drupal-' . end($class) . '-');
-        if (!$tmpDir) {
-            throw new \Exception();
-        }
-
-        unlink($tmpDir);
-        mkdir($tmpDir);
-        $this->tmpDirs[] = $tmpDir;
-
-        return $tmpDir;
     }
 }
